+++ packages/public.config/src/cjs.config.ts
@@ -1,0 +1,4 @@
+import './config.js';
+import init from './cjs.init.js';
+
+init();