+++ packages/public.config/src/cjs.main.ts
@@ -1,0 +1,3 @@
+import init from './cjs.init.js';
+
+init();