+++ packages/public.config/src/configs.ts
@@ -1,0 +1,5 @@
+export type Env = typeof process.env;
+
+const configs = [process.env, (globalThis as unknown as { ENV: Env }).ENV];
+
+export default configs;