+++ packages/public.config/src/esm.config.ts
@@ -1,0 +1,4 @@
+import './config.js';
+import init from './esm.init.js';
+
+init();