+++ packages/public.config/src/esm.configs.ts
@@ -1,0 +1,5 @@
+import configs, { Env } from './configs.js';
+
+configs.splice(-1, 0, (import.meta as unknown as { env: Env }).env);
+
+export default configs;