+++ packages/public.config/src/esm.init.ts
@@ -1,0 +1,6 @@
+import configs from './esm.configs.js';
+import createInit from './createInit.js';
+
+const init = createInit(configs);
+
+export default init;