+++ packages/public.config/src/esm.loadRuntimeConfig.ts
@@ -1,0 +1,6 @@
+import createLoadRuntimeConfig from './createLoadRuntimeConfig.js';
+import init from './esm.init.js';
+
+const loadRuntimeConfig = createLoadRuntimeConfig(init);
+
+export default loadRuntimeConfig;