+++ packages/public.config/src/esm.main.ts
@@ -1,0 +1,3 @@
+import init from './esm.init.js';
+
+init();