+++ packages/react-app-super-template/src/keycloak.ts
@@ -1,4 +1,3 @@
-import config from '@visma/public.config';
 import Keycloak from 'keycloak-js';
 
-export default Keycloak(config.keycloak);
+export default Keycloak(globalThis.ENV?.KEYCLOAK);