+++ packages/react-app-super-template/src/types/index.d.ts
@@ -1,8 +1,8 @@
-declare module 'api' {
-  export const clientPromise: Promise<any>;
+declare module '@/api' {
+  export const clientPromise: Promise<import('axios').AxiosStatic>;
 }
 
-declare module 'api/schema.json' {
-  const _default: any;
+declare module '@/api/schema.json' {
+  const _default: import('openapi-types').OpenAPIV3.Document;
   export default _default;
 }