+++ packages/react-app-super-template/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.0
+
+- **Breaking:** Use project alias `@` to import api schema
+- **Breaking:** Upgrade to `@visma/public.config@1.0.0`
+- Add props to AppProvider for customization
+
 # 0.1.5
 
 - Upgrade to `@visma/formatjs-scripts@0.2.0`