+++ packages/vite-plugin-super-template/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.1
+
+- Set dev server port to 3000
+- Set project resolve alias `@` to `src`
+- Add `@esbuild-plugins/node-globals-polyfill`
+- Add `vite-tsconfig-paths`
+- Convert `src` to TypeScript
+
 # 0.1.0
 
 - Initial release